+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - ** Destination connector adding**: MongoDB
 
+- ** Source connector addding**: OneDrive
+
 - List all the tools in a table.
 
 - Capability to only log request parameters to UnstructuredClient's AsyncHttpClient, no error response is logged by this mechanism.